Moves software keys to file
