+++ src/App.svelte
@@ -10,6 +10,7 @@
 	import UnlockWallet from "./view/UnlockWallet.svelte";
 	import CreateWallet from "./view/CreateWallet.svelte";
 	import {walletDataExists} from "./machinery/secure-storage";
+	import SoftwareKeys from "./view/SoftwareKeys.svelte";
 
 	let header: string | undefined = undefined
 	let view: string | undefined = undefined
@@ -21,15 +22,12 @@
 	});
 
 	onMount(() => {
-		if(walletDataExists()) {
+		if (walletDataExists()) {
 			pushMenu('unlock')
 		} else {
 			pushMenu('create')
 		}
 	})
-
-	const toggleMenu = () => pushMenu('menu')
-
 </script>
 
 <main>
@@ -48,10 +46,5 @@
 			<CreateWallet />
 		{/if}
 	</div>
-
-	<div class="kui-software-key">
-		<h5 role="button" class="kui-h5 kui-text-left" id="kui-left-soft-key"></h5>
-		<h5 role="button" class="kui-h5 kui-text-center kui-text-upcase" id="kui-middle-soft-key"></h5>
-		<h5 role="button" class="kui-h5 kui-text-right" on:click={toggleMenu} data-l10n-id="rightNavButton"></h5>
-	</div>
+	<SoftwareKeys/>
 </main>
